refactor: LLM formatters for OpenAI Responses streaming support
  - add bidirectional OpenAI chat⇄responses formatters that preserve reasoning/thinking metadata
  - introduce streaming accumulators and utilities to rebuild complete responses from chunked events
  - expand provider/config models, CLI flows, and docs to surface the new response pipeline
  - refresh fixtures and tests for OpenAI, Copilot, and Anthropic integrations to cover the new formatters

+++ ccproxy/auth/oauth/flows.py
@@ -111,7 +111,7 @@ class CLICallbackServer:
         return aiohttp.web.Response(text=html_content, content_type="text/html")
 
     async def wait_for_callback(
-        self, expected_state: str | None = None, timeout: int = 300
+        self, expected_state: str | None = None, timeout: float = 300
     ) -> dict[str, Any]:
         """Wait for OAuth callback with optional state validation.
